Dnsmasq displayed as .conf file format
Using dnsmasq as auth. DNS server is more often with a .conf file to load the zones, then by appending all zone file contents as parameters to the executable. Not all users have enough background knowledge to understand that '--dns-rr' is equivalent to 'dns-rr' in the dnsmasq.conf file according to https://serverfault.com/questions/914108/how-to-add-a-caa-record-to-dnsmasq/915012#915012. Therefore show the most often used form: without -- prefix.